requirements.txt
python-telegram-bot==22.3
fastapi==0.112.*
uvicorn==0.30.*
